trainer/requirements.txt
tensorflow
jupyter
azureml-sdk[notebooks]
yaml
-e git+https://github.com/qubvel/efficientnet@e9fdd43857785fe5ccf3863915dcaf618b86849f#egg=efficientnet
